Moved title from AbstractEntity to Item class so that User class can extend from AE.
